+++ common/src/main/java/fr/loudo/narrativecraft/network/handlers/ServerPacketHandler.java
@@ -472,7 +472,8 @@ public class ServerPacketHandler {
                 NarrativeCraftMod.getInstance().getCharacterManager().resolveCharacter(characterId, scene);
         if (characterStory == null) return;
 
-        CharacterPlacement placement = new CharacterPlacement(characterStory, position, rotation, items);
+        CharacterPlacement placement =
+                new CharacterPlacement(characterStory, position, rotation, items, player.onGround());
         String placementJson = CameraAngleSerializer.serializeSingleCharacterPlacement(placement);
         Services.PACKET.sendToPlayer(
                 (ServerPlayer) player, new S2CCameraAngleCharacterCaptured(cameraAngle.getId(), placementJson));
@@ -508,7 +509,8 @@ public class ServerPacketHandler {
             if (!stack.isEmpty()) items.add(stack.copy());
         }
 
-        CharacterPlacement placement = new CharacterPlacement(characterStory, position, rotation, items);
+        CharacterPlacement placement =
+                new CharacterPlacement(characterStory, position, rotation, items, player.onGround());
         String placementJson = CameraAngleSerializer.serializeSingleCharacterPlacement(placement);
         Services.PACKET.sendToPlayer(
                 (ServerPlayer) player, new S2CCameraAngleCharacterCaptured(mainScreenAngle.getId(), placementJson));