+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Fixes
 
 - World locked if leaving the world while a playback is playing
+- Character placement on ground in camera angle not accurate